Split Edit window in two tabs Network/Properties
We already had a tab a little further down for options, but that was
pretty empty. Now moved the IDString prominent to the Network page.
Thisway is on the Network page all info we can see in the Networklist
and on the new Properties tab all config settings.
The advantage is that the window takes up less space in height, which
should be advantageous for small screens.

Add new auth options
  - AUTH_WPA3_EAP for WPA3 Enterprise
  - AUTH_IEEE8021X_WIRED not sure if sensible
both are more a reminder, not really implemented