Disable the new added key_mgmt box for now
Not sure if this new added box make sense.
On the other hand, what sense does the pairwise box make?

As the fog is slowly clearing for me with regard to these many
theoretical possibilities, I think it would be better to remove
some fields and make more auto dependent on the authbox.