Ignore html files in doc/ for now
To have some html help file available from the GUI would be nice,
but ATM I'm not in the mood to look closer at this topic.

At the end of the day, you don't really need a help page at your
finger tips. Documentation is important, but it's more than enough
to have it as man page. You look at it once, then probably
never again.